delete toolWindowMap in StreamingLogsToolWindowManager
