+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib/src/main/java/com/microsoft/azure/toolkit/intellij/common/StreamingLogsToolWindowFactory.java
@@ -9,6 +9,10 @@ import com.intellij.icons.AllIcons;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowFactory;
+import com.intellij.ui.content.ContentManagerEvent;
+import com.intellij.ui.content.ContentManagerListener;
+import com.microsoft.azure.toolkit.lib.common.bundle.AzureString;
+import com.microsoft.azure.toolkit.lib.common.messager.AzureMessager;
 import org.jetbrains.annotations.NotNull;
 
 public class StreamingLogsToolWindowFactory implements ToolWindowFactory {
@@ -26,5 +30,23 @@ public class StreamingLogsToolWindowFactory implements ToolWindowFactory {
     @Override
     public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow toolWindow) {
         toolWindow.setIcon(AllIcons.Nodes.LogFolder);
+        toolWindow.getContentManager().addContentManagerListener(new ContentManagerListener() {
+            @Override
+            public void contentRemoved(ContentManagerEvent contentManagerEvent) {
+                final String displayName = contentManagerEvent.getContent().getDisplayName();
+                StreamingLogsToolWindowManager.getInstance().removeConsoleViewName(displayName);
+            }
+
+            @Override
+            public void contentRemoveQuery(@NotNull ContentManagerEvent event) {
+                final String displayName = event.getContent().getDisplayName();
+                final boolean canClose = AzureMessager.getMessager().confirm(AzureString.format(
+                        "This will stop streaming log of \"{0}\", are you sure to do this?", displayName));
+                if (!canClose) {
+                    event.consume();
+                }
+            }
+
+        });
     }
 }
